Text display + new_line + some basic outputs for kernel entry
bootloader, goes into the kernel at 0x10000, displays some text. custom 4x6 font, but to account for spcaing it increments by 5 on the y. so 320/5 = 64 chars by 200/6 = 33.33 so up to 198 is then 64chars by 33 for printing text with my text to 13h mode. that is 64x33 chars = 2112, which even though its smaller characters is more than the vga text mode 2000chars per page. Next steps are to build a shell, where it takes whatever input into input buffer, then that input will be parsed later on with a parser to create command/language interactions with the shell.