Shifter label working. Fix in shift_screen_4 required. Implement backspace when move_left>0 required.
Shifer label works. Has 3 functions where they, shift pixel data from cursorx,y to 0xAFFFF(end of 13h memory), the other shifts all bytes in page_data by 1 containing data of page for each character 2112bytes, and the last shifts input buffer from the index of x and y offset. of cursor All these shift from the cursor_offsetx and cursor_offsety. Currently a value can be entered and edited with new text inside the input, but the backspace still only deletes from the end.

implement: backspace when move_left > 0 meaning inside of the input buffer not at the end

fix: cursor_offsetx > 315, or its on the last character, and move_left > 0, and i try to shift here, it shifts so wrong when teh cursor is at that position. Either fix or add a limitation. The error lies in how the pixel are visually shifted so the error lies somewhere in shift_screen_4