Implemented one keyword, new_file(), opens kernel_4.0_newfile.asm, boots in protected mode now
newfile opens just another mainloop without the kernel, and one large input area.
limited to 32 rows for now, implemented shifter_right so that i dont have to worry about the input buffer .
need to implement a way to escape from this new_file(). 

Protected mode I implemented a flat memory model of 1 code segment and 1 data segment from 0x00000 to 0xfffff so 1mb of space to work with for now.
There is also as stack ;one 4096 byte area at 0x100000
Then in index 4 and 5 of the gdt are spaces within the code and data segemnt area, just defining a base so that they can be easily indexing the 32 bit base in this case index 4 being 0xA0000 to 0xFFFFF this is so that 0x20 points to video ram, and then 0x28 points to 0xCFFFF my final space that spans from 0xCFFFF-0xFFFF in which i will use to implement some paging space for my files later on.