Add files via upload
print_hex_to_decimal was implemented.
takes value at locaiton data label/location passed to esi and prints it as a string but the deimcal representation.
Loops from 10^9 down to 10^0 for each digit of possible 10 decimal digits for 32 bit mode where hex will have max 8 hex digits and 10 decimal degits.
In the case that the value is smaller such as 0x00000009, when dividing by 10^9 down, it looks for the first non zero char. thus it will print 9, but 0x00000064(100 decimal)  will print 100 not 0000000100. thus ignoring all zeros up to the first non zero, and in the case that it is a 0 it just prints 0 at the end before return. This is called in data to number the lines.


implement/fix:
-fix line numbering to not be deletable in a new_file()
-make sure when line wrapping occurs to still show the next line number or it will skip
-can implement a label that holds the offset of the print of each line and controls when the cursor skips back or deletes by skipping over when x_offset is equal to each lines number offset. ex 1. would be an offset of 2 cause 2 character. 100. would be an offset of 4 becaue that is 4 chars to print "100."

assembler file initialized and created a keyword for a testing area for this. Goal is to assemble code from one label that is x86 assembly, and move the machine assembled code into another label.