parsinging label parses by spaces. grabbing into 3 buffers for each instruction and then resetting for next instruction. + fixed (still required edits) to print hex to decimal.
opcode, register mode tables added etc for indexing commands. 
enable bits for irm to r, or r to rm, or  im32 to rm, or im32 to r. If immediate value found turns to 1, if a bracket is found its r, r/m (register to register/memory)
currently the first instruction so mov, add, sub, cmp or error if first term is anything else, is indexed into opcode tables and written into output_machine_code from input_asm label as hex machine code.  

print_hex_as_decimal broke somehow so i worked on various other versions.
i came up with print_hex_as_decimal3 in which i only go for byte at a time, this can be changed at the line of movzx eax, byte [esi] to work with grabbing larger hexes from labels. Thus I can currently only print out values up to 255. This was more of a design choice in the moment because if i grabbed larger amounts than a byte from [esi] I was grabbing some junk when working with machine_code_output and wanting to view my byte outputs. So i grab as a byte and extend to 32 bits for now. Meaning at any point I can change it to grab more bytes from that line and it will work from any hex value in a register up to 32 bits. This is fine for now but later if I want to vosualize hex values larger than 255 I need to tweak the code or I can just change it to have some kind of enable for the amount of bytes being taken at position esi.


There was also another bug in print_hex_as_decimal in which i was using esi for the loop but tweaking the actual position, which even though i was tracking the output_machine_code elsewhere and i preserved all registers, because it was tied to it, it still edited the original values after its called, which was confusing me on why my output_machine_code output was dissapearing randomly. Eventually i found this bug and I preserve the value at esi when print hex is called so it doesnt overwrite.