fixed print_hex_to_decimal3 to use print_hex_to_decimal_3_helper to avoid overwriting value at [esi].
The reason this change was necessary was becasue if i wanted to print a hex value with print_hex_to_decimal3 after printing the value at [esi] would be overwritten. The issue lied in this function so I had to give it another value that held the same hex as in esi so I copy it into print_hex_to_deimal_3_helper as i do the divisions and change [print_hex_to_decimal_3_helper] as i go allowing whatever hex i wanted to print out to stay in esi rather than being overwritten.