parsing of keywords for colour done. example somehting(0x123128) will return something into value1buffer and 123128 hex into value2buffer, if theres an error both will be empty
parses looking for first (, then on second bracket close ) when appending to term2 buffer, it will limit the digits to be 8 digits excluding the 0x which gets parsed off. Example 0x12345678 is fine but 0x123456789 will be too large to enter as a hex value.   Next steps: compare the first term value with keywords, and apply the hex if the hex passed is <255, then take the lower byte and change the bg colour or text and bg color or text revert to change the bg colour or text colour on keywords entered.