template for LBA129 to be used for colour saves
plan to update the colour change data on set_bg() and set_text() changes, scan the 512 byte long data label looking for a 't' or 'b' depending on if the text or bg was changed, if found, just update the next byte to the colour changed, if not found, write the ascii +the colour byte to save. Then when reading from LBA129, if 't' or 'b' is found, the next byte can be fed directly to bg_colour, bg_revert, text_colour, text_revert depending on whats found in the saved data in that sector.