write_0xff_tosave: created and LBA129_write_coloursaves: create
write to the 512 bytes was created LBA129_savedata, it is initialized as 0xff, when printall_colours() is called, write_0xfftosave is called in which it reads whats in the save data into the disk at sector LBA 129. This is tested by when the code boots, LBA129_savedata is filled with zeros from disk, then when command is run, close the program, run solely qemu-system-i386 -hda final_version2.img not the entire batch file, and then the disk saves on this disk image will be contained from that 129 LBA sector thus on the next boot at start when printing savedata the saved 0xff's will be printed as 255's rather than as zeros this time showing the save worked.