implemented saving bg and text colours in LBA129(512 bytes) and pulling that data on kernel starti=up into data label LBA129_savedata
using check_colours when set_bg and set_text are called will look to see if bytes 'b' and 't' exists and if not write that byte and the new colour in the next byte, if it exists just write the new byte next. On startup this is called again to set the changed colours directly from LBA129_savedata into bg and text colour to overwrite the default when qemu-system-i386 -hda final_version2.img is called, which is just running only the img file, if you call kernel_testing_batch2.bat the bootloader resets the whole thing and it all starts again, but the img file itself keeps the save cause its a disk save.